+++ src/tests/contract/getById.test.js
@@ -1,7 +1,7 @@
 const request = require("supertest");
-const app = require("./app.js");
-const { seed } = require('../scripts/seedDb');
-const { HTTP_STATUS_CODES, stripTimestamps } = require('./common');
+const app = require("../../app.js");
+const { seed } = require('../../../scripts/seedDb');
+const { HTTP_STATUS_CODES, stripTimestamps } = require('../../common');
 
 describe('Get contract by Id', () => {